refactor(breadcrumbs): combine useParams and useLocation for path generation
