--- .eslintcache
@@ -1,1 +1,0 @@
-[{"C:\\Users\\sobse\\OneDrive\\Documents\\rs-school\\final-task\\CodeMind\\src\\components\\breadcrumbs\\breadcrumbs.tsx":"1","C:\\Users\\sobse\\OneDrive\\Documents\\rs-school\\final-task\\CodeMind\\src\\components\\header\\header.tsx":"2","C:\\Users\\sobse\\OneDrive\\Documents\\rs-school\\final-task\\CodeMind\\src\\utils\\crumbs-path.ts":"3"},{"size":777,"mtime":1772961280334,"results":"4","hashOfConfig":"5"},{"size":1856,"mtime":1772960839012,"results":"6","hashOfConfig":"5"},{"size":856,"mtime":1772962220497,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","suppressedMessages":"10","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1wk39hh",{"filePath":"11","messages":"12","suppressedMessages":"13","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"14","messages":"15","suppressedMessages":"16","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"C:\\Users\\sobse\\OneDrive\\Documents\\rs-school\\final-task\\CodeMind\\src\\components\\breadcrumbs\\breadcrumbs.tsx",[],[],"C:\\Users\\sobse\\OneDrive\\Documents\\rs-school\\final-task\\CodeMind\\src\\components\\header\\header.tsx",[],[],"C:\\Users\\sobse\\OneDrive\\Documents\\rs-school\\final-task\\CodeMind\\src\\utils\\crumbs-path.ts",[],[]]
