+++ src/components/breadcrumbs/breadcrumbs.tsx
@@ -1,20 +1,22 @@
 import { NavLink, useLocation } from 'react-router-dom';
 import styles from './breadcrumbs.module.css';
-import { createCrumbsPath } from '../../utils/crumbs-path';
+
 import { Fragment } from 'react/jsx-runtime';
+import { Paths } from '../../constants';
+import { useBreadcrumbs } from './hooks/use-breadcrumbs';
 
 export const HeaderBreadcrumbs = () => {
   const { pathname } = useLocation();
-  const crumbs = createCrumbsPath(pathname);
+  const crumbs = useBreadcrumbs();
   return (
     <>
       {crumbs.map(({ name, path }, index) => {
         return (
-          <Fragment key={index}>
+          <Fragment key={path}>
             {index === 0 ? '' : <span> /</span>}
             <NavLink
               to={path}
-              className={`${styles.link}  ${pathname === path || crumbs.length === 1 ? styles.active : ''}`}
+              className={`${styles.link}  ${pathname === path || pathname === `${Paths.COURSES}/${Paths.MY_COURSES}` || pathname === `${Paths.COURSES}/${Paths.ALL_COURSES}` ? styles.active : ''}`}
             >
               {name}
             </NavLink>
